Regroup Diagnostics into Settings/Diagnostics/Research tabs, move Worktrees to Usage Analysis
The Diagnostics view had grown to 12 flat tabs mixing settings, troubleshooting,
and analytical research together. Split it into three top-level groups with
their own leaf-tab bar: Diagnostics (Report, Session Files, Cache, Path
Analyzer), Research (Model Usage, Tool Analysis, OTel Delta), and Settings
(Display, Backend Storage, GitHub Auth, Debug).

The Worktrees tab is a repo-maintenance tool, not diagnostics or research, so
it moves to the Usage Analysis view as its own tab instead. Backend message
handling and the live worktree-root discovery push were rewired from the
diagnostics panel to the analysis panel accordingly.
