+++ vscode-extension/src/webview/diagnostics/styles.css
@@ -79,7 +79,7 @@ body {
 	margin-bottom: 16px;
 }
 
-.tab {
+.tab, .group-tab {
 	padding: 10px 20px;
 	cursor: pointer;
 	border: none;
@@ -91,12 +91,12 @@ body {
 	transition: all 0.2s;
 }
 
-.tab:hover {
+.tab:hover, .group-tab:hover {
 	color: var(--text-primary);
 	background: var(--list-hover-bg);
 }
 
-.tab.active {
+.tab.active, .group-tab.active {
 	color: var(--link-color);
 	border-bottom-color: var(--link-color);
 }
@@ -107,6 +107,21 @@ body {
 
 .tab-content.active {
 	display: block;
+}
+
+/* Group tab bar (Diagnostics / Research / Settings) sits above the leaf tab bar for its group */
+.group-tabs {
+	border-bottom: 2px solid var(--border-color);
+	margin-bottom: 4px;
+}
+
+.group-tab {
+	font-size: 14px;
+	font-weight: 600;
+}
+
+.leaf-tabs {
+	margin-bottom: 16px;
 }
 
 /* Sub-tabs (used inside tabs, e.g. Backend Storage sub-tabs) */
@@ -836,172 +851,4 @@ border: 1px solid #1f6feb;
 @keyframes spin {
 	to { transform: rotate(360deg); }
 }
-
-/* Worktrees tab */
-.worktree-roots-list {
-	display: flex;
-	flex-direction: column;
-	gap: 6px;
-}
-
-.worktree-root-item {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	gap: 8px;
-	background: var(--bg-tertiary);
-	border: 1px solid var(--border-color);
-	border-radius: 4px;
-	padding: 6px 10px;
-	font-family: var(--vscode-editor-font-family, monospace);
-	font-size: 12px;
-}
-
-.worktree-root-item span {
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.worktree-progress-bar {
-	height: 6px;
-	border-radius: 3px;
-	background: var(--bg-tertiary);
-	overflow: hidden;
-	margin-top: 8px;
-}
-
-.worktree-progress-fill {
-	height: 100%;
-	background: var(--link-color);
-	transition: width 0.2s ease;
-}
-
-/* While walking the folder tree we have no percentage yet, so pulse the bar to
-   signal ongoing activity instead of showing a misleading fixed progress. */
-.worktree-progress-fill.indeterminate {
-	animation: worktree-pulse 1.2s ease-in-out infinite;
-}
-
-@keyframes worktree-pulse {
-	0%,
-	100% {
-		opacity: 0.35;
-	}
-
-	50% {
-		opacity: 1;
-	}
-}
-
-.worktree-repo-row {
-	cursor: pointer;
-	font-weight: 600;
-}
-
-.worktree-repo-row:hover {
-	background: var(--bg-tertiary);
-}
-
-.worktree-repo-row.expanded {
-	background: var(--bg-tertiary);
-}
-
-.worktree-delete-link {
-	margin-left: 8px;
-	color: var(--vscode-errorForeground, #f14c4c);
-}
-
-.worktree-delete-link:hover {
-	text-decoration: underline;
-}
-
-.worktree-caret {
-	display: inline-block;
-	width: 14px;
-	color: var(--text-muted);
-	font-size: 10px;
-}
-
-.worktree-roots-toggle {
-	display: inline-flex;
-	align-items: center;
-	gap: 4px;
-	margin: 8px 0;
-	padding: 0;
-	background: none;
-	border: none;
-	color: var(--link-color);
-	font-size: 12px;
-	cursor: pointer;
-}
-
-.worktree-roots-toggle:hover {
-	text-decoration: underline;
-}
-
-.worktree-pending {
-	color: var(--text-muted);
-	font-style: italic;
-	opacity: 0.8;
-}
-
-/* The details row's cell wraps the per-worktree table; trim its padding so the
-   nested table aligns with the parent columns. */
-.worktree-repo-details > td {
-	padding: 0 0 12px 0;
-}
-
-.worktree-cleanup-card {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	gap: 6px;
-}
-
-.worktree-cleanup-card-actions {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	gap: 6px;
-	flex-wrap: wrap;
-}
-
-.worktree-cleanup-card-actions .button {
-	font-size: 12px;
-	padding: 4px 10px;
-}
-
-.worktree-cleanup-log {
-	margin-top: 8px;
-	display: flex;
-	flex-direction: column;
-	gap: 4px;
-	font-size: 12px;
-}
-
-.worktree-cleanup-log-row {
-	display: flex;
-	gap: 8px;
-	align-items: baseline;
-	padding: 4px 6px;
-	border-radius: 4px;
-	background: var(--bg-tertiary);
-}
-
-.worktree-cleanup-log-branch {
-	font-weight: 600;
-	font-family: var(--vscode-editor-font-family, monospace);
-}
-
-.worktree-cleanup-log-repo {
-	color: var(--text-muted);
-}
-
-.worktree-cleanup-log-reason {
-	color: var(--text-muted);
-	flex: 1;
-}
-
